requirements.txt
jupyter
pandas
scikit-learn
xgboost
lightgbm
numpy
matplotlib
seaborn
optuna
hydra-core
omegaconf
typer[all]
loguru
mlflow
gradio
joblib
pyyaml
black
isort
ruff
pytest